piegen: fix build on 64-bit ubuntu
The error I got was:

  CC       pie/piegen/elf-x86-64.o
In file included from pie/piegen/elf-x86-32.c:16:0:
pie/piegen/elf.c: In function ‘handle_elf_x86_32’:
pie/piegen/elf.c:476:3: error: format ‘%lx’ expects argument of type ‘long unsigned int’, but argument 6 has type ‘Elf32_Word’ [-Werror=format=]
   pr_debug("Copying section '%s'\n" \
   ^

+++ pie/piegen/elf.c
@@ -476,7 +476,7 @@ int handle_elf(void *mem, size_t size)
 		pr_debug("Copying section '%s'\n" \
 			 "\tstart:0x%lx (gap:0x%lx) size:0x%lx\n",
 			 &secstrings[sh->sh_name], (unsigned long) sh->sh_addr,
-			 (unsigned long)(sh->sh_addr - k), sh->sh_size);
+			 (unsigned long)(sh->sh_addr - k), (unsigned long) sh->sh_size);
 
 		/* write 0 in the gap between the 2 sections */
 		for (;k < sh->sh_addr; k++) {
